[Backend] Add new version JDM-Legends Commons

+++ JDM-Legends/JDM-Legends-Commons/src/main/java/com/jdm/legends/common/dto/Car.java
@@ -47,7 +47,7 @@ public class Car {
 
     private int productionYear;
 
-    private BigDecimal initialPrice;
+    private Integer initialPrice;
 
     private int hp;
